+++ test/test_source/test_inc.f90
@@ -10,5 +10,6 @@ SUBROUTINE test_sub
 
 val2
 END SUBROUTINE test_sub
+include 'mpi.f'
 
 END MODULE test_mod